add clear_file and clear_all_files
